+++ lib/apiScenario/apiScenarioTypes.ts
@@ -136,6 +136,7 @@ export type StepRestCall = StepBase & {
   parameters: SwaggerExample["parameters"];
   responses: SwaggerExample["responses"];
   outputVariables?: OutputVariables;
+  externalReference?: boolean;
   _resolvedParameters?: SwaggerExample["parameters"];
 };