+++ ChangeLog.md
@@ -7,6 +7,7 @@
   - Support readme.test.md run API test
   - Fix step variable unresolved in newman collection
   - Fix bugs about html report
+  - Aggregate reports into one per scenario file
 
 ## 07/06/2022 3.0.3
 - Generate high quality examples from API Scenario tests